requirements.txt
fastapi
datasets
pillow
torch
transformers
fastapi[standard]